requirements.txt
Flask==0.11.1
Flask-SQLAlchemy==2.3.2
Flask-Limiter==1.0.1
itsdangerous==0.24